Revert some change in order forms model

+++ order/api/src/test/kotlin/com/rarible/protocol/order/api/service/order/AbstractOrderIt.kt
@@ -138,7 +138,7 @@ abstract class AbstractOrderIt : AbstractIntegrationTest() {
             AssetConverter.convert(make),
             taker,
             AssetConverter.convert(take),
-            salt.toBigInteger(),
+            salt,
             start,
             end,
             OrderDataConverter.convert(data),
